+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 1. Add initial nextjs setup
 2. Integrate next-auth
+3. Integrate discord auth